# Add 8x4 Arm64 kernel to improve performance