In my spare time I'm been messing around with the optimised Arm64 kernels.

This PR adds an 8x4 kernel (for 64-bit bitpacked input) which improves single-threaded benchmark performance by around 6%.

## What do these changes do?

The current optimal Arm kernel (for 64-bit bitpacking) has a 4x4 layout and uses 32-bit accumulators. This PR additionally introduces an 8x4 version of this kernel that uses 16-bit accumulators.

After the accumulation loop, the first operation is an `unsigned shift left long` instruction which performs the necessary back-transformation left-shift by 1, and writes the result into 32-bit vector registers. This means that all subsequent 'post-processing' (clamping, post-multiply and post-addition) is done in Int32/Float32, as in the existing kernel.

### Performance improvements

Performance is improved (compared to the 4x4 kernel) in two ways:

* **Increased weight sharing**: The 8x4 layout means that the RHS data gets used two times more for each load, or equivalently the number of RHS memory loads is halved.

* **Faster BMLA pseudo-instructions**: With 8 rows and 16-bit accumulators, each column of accumulators fits into a single NEON register. With 16-bit accumulators rather than 32-bit accumulators, there's no need for an extra `uadalp` instruction to pack 16-bit results into 32-bits (specifically, there's no need for the equivalent of [this instruction](https://github.com/larq/compute-engine/blob/f801c05b9373528b29b8732ae55e3cc605d26747/larq_compute_engine/core/bgemm_kernels_arm64.h#L25) in the current kernel). As a result, the proposed `BMLA` macro uses 24 instructions to process eight rows of data, as opposed to 13 instruction to process four rows of data in the current 4x4 kernel.

### Overflow risk with 16-bit accumulators

There is a potential risk of overflowing the 16-bit accumulators if `filter_height` * `filter_width` * `channels_in` >= 2^16. As a result, this PR does not replace the existing 4x4 kernel, and safely falls back to the old kernel if this overflow threshold is breached.

In reality, we will always use the new 8x4 kernel for models of any reasonable size: a `3x3` kernel would have to have > 7000 input channels to present an overflow risk.

## How Has This Been Tested?

All existing kernel tests pass, including the 'big' versions of the tests (that I've run locally, as they don't run by default on CI).

## Benchmark Results

I have benchmarked current master (the 'Baseline') and this PR on my Android phone (a OnePlus 6) and on a Raspberry Pi Model 4B. I used the LCE benchmark utility with `num_runs=250` and report the average and the standard-deviation.

#### Single-threaded

| Device    | Model         | Baseline      | PR            | Speed-up |
|-----------|---------------|---------------|---------------|----------|
| OnePlus 6 | QuickNet      | 19.83 +- 0.28 | 18.72 +- 0.21 | 5.6%     |
| OnePlus 6 | QuickNetLarge | 28.41 +- 0.22 | 26.99 +- 0.22 | 5.0%     |
| OnePlus 6 | QuickNetXL    | 50.65 +- 0.52 | 47.86 +- 0.33 | 5.5%     |
| RPi 4     | QuickNet      | 37.82 +- 0.09 | 34.94 +- 0.17 | 7.6%     |
| RPi 4     | QuickNetLarge | 55.98 +- 0.12 | 52.66 +- 0.15 | 5.9%     |
| RPi 4     | QuickNetXL    | 99.09 +- 0.46 | 92.00 +- 0.26 | 7.2%     |

#### Multi-threaded (4 threads)

| Device    | Model         | Baseline      | PR            | Speed-up |
|-----------|---------------|---------------|---------------|----------|
| OnePlus 6 | QuickNet      | 10.36 +- 0.40 | 10.08 +- 0.31 | 2.7%     |
| OnePlus 6 | QuickNetLarge | 14.65 +- 0.55 | 14.14 +- 0.42 | 3.5%     |
| OnePlus 6 | QuickNetXL    | 26.32 +- 0.53 | 25.34 +- 0.55 | 3.7%     |
| RPi 4     | QuickNet      | 17.73 +- 0.34 | 17.17 +- 0.45 | 3.2%     |
| RPi 4     | QuickNetLarge | 26.56 +- 0.56 | 25.74 +- 0.43 | 3.1%     |
| RPi 4     | QuickNetXL    | 43.63 +- 1.09 | 41.71 +- 1.06 | 4.4%     |

There is an average single-threaded speed-up of 6.1% and multi-threaded speed-up of 3.4%.