@AdamHillier 2020-05-09
### Prospects for an 8x8 kernel

If we don't use `RUY_OPT_MAX_STREAMING`, then there are enough registers to be able to implement an 8x8 kernel. I have experimented with disabling the `RUY_OPT_MAX_STREAMING` part of this kernel and haven't noticed any significant benchmarked performance difference, so I don't think that would be a problem.

However, I also don't think that going to 8x8 would bring significant futher performance gains in and of itself, because as far as I can tell most of the performance gain in this PR comes from the slightly more optimal `LCE_BMLA` rather than the increased weight sharing. That's just a guess though.

On [this branch](https://github.com/AdamHillier/compute-engine/tree/8x8-arm64-kernel) I have an implementation of an 8x8 kernel that is very close to working (out of the many thousands of kernel tests, it fails 12 of them - specifically it fails for large inputs combined with valid padding combined with large stride combined with large dilation). I haven't been able to successfully debug the issue, and I think @Tombana also had weird specific test failures when he tried to get an 8-column kernel working, so there might be some weird underlying issue. In any case, I'm not going to pursue 8x8 any further any time soon.