@arashb 2020-05-09
> ### Prospects for an 8x8 kernel
> If we don't use `RUY_OPT_MAX_STREAMING`, then there are enough registers to be able to implement an 8x8 kernel. I have experimented with disabling the `RUY_OPT_MAX_STREAMING` part of this kernel and haven't noticed any significant benchmarked performance difference, so I don't think that would be a problem.

I observed the same performance behaviour, utilizing the entire register space did not yield noticeable performance differences (it decreases the iteration over depth and as a result number of kernel calls but this is not significant)

Regarding the 8x8 kernels, I also think it will not also yield signifant improvement (for the same reason as stated above for increasing usage of register space) However, if you are still interested to get it working, I have a 8x8 kernel working [here](https://github.com/larq/compute-engine/blob/c6d488a424f02aa97b4fdfd727d26e9033afe50b/larq_compute_engine/tflite/cc/kernels/bgemm_kernels_arm64.h#L45) (with a completely different memory layout and instructions). You might have a bug in corner case where the destination matrix does not fit entirely into 8x8 sub-matrices. You can check against my branch to debug that.