@AdamHillier 2020-05-10
We previously had a separate 4x4 kernel for uint32 inputs (that didn't, for example, implement `RUY_OPT_MAX_STREAMING`). As far as I can tell, there's no reason not to write the optimised kernels to just accept uint64, i.e. get rid of the separate uint32 kernel, because we can still use the uint64 kernels with uint32 inputs (by pretending that the inputs are uint64: halving the LHS/RHS stride and the depth in the kernel params).

That is what I have implemented in ab533ea. All the big tests, in particular the `OPT_32` tests, still pass.

The benefit of doing this is that we remove a bunch of code and reduce binary size. Before this PR there were two Arm64 kernels: one 4x4 uint32 kernel and one 4x4 uint64 kernel. After this PR there will still be two Arm64 kernels: one 8x4 uint64 kernel and one 4x4 uint64 kernel. But both will be compatible with uint32 input as well. It also means that using 32-bit bitpacked input still allows us to take advantage of the fastest and most optimised kernels, which is important e.g. for reading bitpacked input directly (which is uint32).

The downside is that ab533ea has made the diff less readable so I'm very happy to revert it and open another PR to discuss those changes.