@Tombana 2020-05-11
> That is what I have implemented in [ab533ea](https://github.com/larq/compute-engine/commit/ab533eab50719f76d09e566cb69b977bbcf32c79). All the big tests, in particular the `OPT_32` tests, still pass.

That sounds great! I think i've measured performance differences in the past between the '32-bit' and '64-bit' kernel, even though they should be the same (just some parameters doubled/halved). Could it be that the '64-bit' kernel causes better memory alignment by requesting blocks that are double the size (but half as many) and therefore a speedup? I guess that's the only reason I can think of for having this distinction at all.