@AdamHillier 2020-05-11
> That sounds great! I think i've measured performance differences in the past between the '32-bit' and '64-bit' kernel, even though they should be the same (just some parameters doubled/halved). Could it be that the '64-bit' kernel causes better memory alignment by requesting blocks that are double the size (but half as many) and therefore a speedup? I guess that's the only reason I can think of for having this distinction at all.

Ah it's quite possible that alignment might make a difference; the only other concrete difference I can think of is that the uint32 kernels previously didn't have `RUY_OPT_MAX_STREAMING`, though as discussed above I don't think that makes a difference really. Also I guess the actual bitpacking step would be quicker for uint64?