@Tombana 2020-05-11
> Ah it's quite possible that alignment might make a difference; the only other concrete difference I can think of is that the uint32 kernels previously didn't have `RUY_OPT_MAX_STREAMING`, though as discussed above I don't think that makes a difference really. Also I guess the actual bitpacking step would be quicker for uint64?

Ah yes the bitpacking is only optimized for uint64. I also didn't realize we didn't have the `RUY_OPT_MAX_STREAMING` in the `uint32` kernel. It might be good to benchmark this stuff, with the bitpacking temporarily commented out.

I'm wondering if it makes sense to keep the 32-bit code path at all. There might be some situations where the 32-bit bitpacking path is faster, when the number of input channels is less than 32, so there's less bitpacking-padding overhead. However I think the assembly kernels always pad it to 128-bits (in the "ruy prepacking step"), so maybe its never faster?