@AdamHillier 2020-05-11
> I'm wondering if it makes sense to keep the 32-bit code path at all. There might be some situations where the 32-bit bitpacking path is faster, when the number of input channels is less than 32, so there's less bitpacking-padding overhead. However I think the assembly kernels always pad it to 128-bits (in the "ruy prepacking step"), so maybe its never faster?

Yes I've been thinking about this too. I think the assembly kernels pad to 128-bits only after im2col, so if we have a 3x3 kernel operating on 24 or 32 input channels, for example, and do bitpacking before im2col then there would be a performance gain to be had.

Also, I was under the impression that if/when we implement writing bitpacked activations for the assembly kernels we would do it with 32-bit bitpacked activations as in the reference kernel? Not 100% sure about that though.