@Tombana 2020-05-11
> Yes I've been thinking about this too. I think the assembly kernels pad to 128-bits only after im2col, so if we have a 3x3 kernel operating on 24 or 32 input channels, for example, and do bitpacking before im2col then there would be a performance gain to be had.

Good point, there's im2col inbetween indeed.

> Also, I was under the impression that if/when we implement writing bitpacked activations for the assembly kernels we would do it with 32-bit bitpacked activations as in the reference kernel? Not 100% sure about that though.

I'm not sure either. Can we use a destination matrix with \#rows padded to multiples of 64 even when the LHS-\#rows does not agree with that? It might get a bit hacky at some point, but i guess it already is when writing bitpacked activations.